# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: >-
    Static website hosting with a light per-site backend. Deploy a site with one
    API call — inline JSON files (LLM-friendly) or a tar.gz/zip archive — and each
    site gets shared JSON state (with atomic ops), append-only collections. Sign-in is a 6-digit email code or Google that
    returns an API key, sent as the
    `X-API-Key` header. Hosted-page sign-in uses the same account but a
    site-scoped session cookie on the site's own address — never an API key.
    Every site lives at its own address, `https://<site>.<handle>.simple-host.app/`
    (its own browser origin); `https://<handle>.simple-host.app/` lists the
    person's public sites. For a brand-new account a site may briefly live at
    `https://<handle>.simple-host.app/<site>/` until its certificate is issued:
    use the `site_url` the API returns, never a composed one.
    Old `<handle>.simple-host.app/<site>/` links redirect to the site's address,
    as do old `sites.simple-host.app/<handle>/<site>/` ones.
    Visitors sign in with Google or an emailed code on the site's own address (a
    sign-in covers that site only); every save from a page needs a signed-in visitor. Agents save with the API key
    (or the connector) through `https://simple-host.app/v1/...`.
  version: 2.0.0
servers:
  - url: https://simple-host.app
...
paths:
...
  /v1/sites/{sitename}/collections/{coll}:
    delete:
      operationId: clearCollection
      summary: Empty a list (owner only)
      description: >-
        Deletes every item in one list, public or private, for good; the list
        keeps its private/public setting. The body must repeat the list's name,
        `{"confirm": "<coll>"}`, or nothing is deleted (400 confirm_required).
        Allowed for the site owner (X-API-Key, connector token, or the owner's own
        visitor session on the site's own address with `X-SH-CSRF: 1`) and the
        Simple Host operator, for moderation. Everyone else gets 404 not_found.
        Visitors only ever append.
      tags: [Collections]
      parameters:
        - { $ref: '#/components/parameters/sitename' }
        - { $ref: '#/components/parameters/coll' }
        - { $ref: '#/components/parameters/csrf' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [confirm], properties: { confirm: { type: string, description: "the list's name again" } }, example: { confirm: guestbook } }
      responses:
        '200':
          description: The list is empty
          content:
            application/json:
              schema: { type: object, properties: { site: { type: string }, collection: { type: string }, deleted: { type: integer, description: how many items were deleted } } }
        '400':
          description: The body does not repeat the list's name (confirm_required)
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: "to empty this list, send {\"confirm\": \"guestbook\"}", code: confirm_required } } }
        '403':
          description: A visitor session without `X-SH-CSRF` (csrf_required)
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: The caller is not the owner or operator (not_found)
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: not found, code: not_found } } }
    post:
      operationId: postSitesBySitenameCollectionsByColl
      summary: Append an item to a collection
      description: >-
        Append-only list (O(1) insert). For signups, RSVPs, submissions.
        Origin-gated. Every save from a page needs a signed-in visitor: a visitor session (`X-SH-CSRF: 1`) on the site's own address — its address `https://<site>.<handle>.simple-host.app/` (same-origin `/v1/sites/{sitename}/...`; only that site answers there), or its domain if it has one. Agents write with the site owner's `X-API-Key` (or the connector signed in as the owner; the platform admin's key writes any site) through the apex `https://simple-host.app/v1/...`; another account's key gets 404 `{"error":"site not found"}` and writes nothing. A site with a connected domain takes page writes only there; writes through its other addresses answer 401 use_custom_domain. WRITE_AUTH_MODE=log/off and allow_anonymous_writes retain their compatibility exceptions. Invalid keys return 401 with {error: "invalid API key", code: "invalid_api_key"}. Each item ≤ 64 KB.
        PRIVATE LISTS (see PUT .../collections/{coll}/privacy) have their own rule: an item is accepted only on the site's own address (its address `https://<site>.<handle>.simple-host.app/`, or its domain if it has one), from a page there, from a visitor signed in there (session cookie + `X-SH-CSRF: 1`); API keys, agents and the connector cannot add to a private list (403 private_visitor_only), and any other address answers 403 private_needs_own_domain (401 use_custom_domain for a site with a connected domain). The item must be ONE JSON object. The server stamps `_submitted_by` (the visitor's verified email) and `_submitted_at` (server time, RFC 3339), overwriting any value the page sent under those keys; the 201 answer is the stored item with the stamps, which is how a page shows the visitor what they sent.
      tags: [Collections]
      parameters:
        - { $ref: '#/components/parameters/sitename' }
        - { $ref: '#/components/parameters/coll' }
        - { $ref: '#/components/parameters/csrf' }
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses:
        '201':
          description: Item appended. The stored item; on a private list `data` carries the server stamps `_submitted_by` and `_submitted_at`.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer, example: 42 }
                  data: { type: object, example: { item: "2 jars", _submitted_by: ana@example.com, _submitted_at: "2026-09-24T18:02:11Z" } }
                  created_at: { type: string, format: date-time }
        '400':
          description: Invalid body (not JSON; on a private list, not one JSON object)
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '413': { $ref: '#/components/responses/TooLarge' }
        '401':
          description: No visitor signed in (visitor_auth_required), invalid account API key (invalid_api_key), or a write through another address to a site that has a connected domain (use_custom_domain, with the domain in `domain`) — such a site saves only on its domain; use the apex (`https://simple-host.app/v1/...`, key) or the domain's own `/v1/` (key or session). Reads stay public. use_custom_domain is never returned for a site that has no domain. A private list answers visitor_auth_required whenever no visitor is signed in on the site's own address.
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string }, code: { type: string }, sign_in: { type: string }, retry: { type: boolean }, domain: { type: string, description: The site's bound custom domain (with code use_custom_domain) } } }
              examples:
                invalidKey: { value: { error: "invalid API key", code: invalid_api_key } }
                signIn: { value: { error: "sign-in required to write", code: visitor_auth_required, sign_in: /v1/auth/oauth/providers, retry: true } }
                useCustomDomain: { value: { error: "this site saves on its own domain", code: use_custom_domain, domain: recipes.brand.com } }
        '403':
          description: Origin/Referer does not match the site; missing `X-SH-CSRF` with a visitor session (csrf_required); or, on a private list, an API key was sent (private_visitor_only) or the request did not come through the site's own address (private_needs_own_domain — private lists accept submissions only on the site's own address).
          content:
            application/json:
              schema: { type: object, properties: { error: { type: string }, code: { type: string, enum: [csrf_required, private_visitor_only, private_needs_own_domain] } } }
              examples:
                csrf: { value: { error: "missing CSRF header", code: csrf_required } }
                privateVisitorOnly: { value: { error: "this list is private: only visitors signed in on the site's own address can add to it; API keys cannot", code: private_visitor_only } }
                privateNeedsOwnDomain: { value: { error: "this list is private: it takes submissions only on the site's own address, from a signed-in visitor", code: private_needs_own_domain } }
        '404':
          description: Site not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      operationId: getSitesBySitenameCollectionsByColl
      summary: List items in a collection
      description: >-
        Newest-first, paginated. A public list is readable by anyone: with no
        `Origin` and no `Referer` (curl, an agent) it is served as is; from a page it
        is Origin-gated to the site's own page; the owner/admin API key also reads it. Writes (POST) from a page need a signed-in visitor (session with X-SH-CSRF) on the site's own address; agents write with any valid account X-API-Key through the apex (a site with a connected domain takes page writes only there — 401 use_custom_domain elsewhere). Origin gating still applies.
        A PRIVATE list is readable only by its owner — and by the Simple Host operator (platform admin key), for moderation: with the owner's `X-API-Key` or connector token (through the apex), or by the owner's own account signed in as a visitor on the site's own address (so an admin page on the site can list it). Each item carries its `id`, which the owner uses to edit or delete it (PATCH/DELETE .../items/{id}). Everyone else — anonymous, other signed-in visitors, other sites, anything through another address — gets 404 `{"error":"not found","code":"not_found"}`. Visitors cannot read back their own submissions. The answer for a private list adds `"private": true`.
      tags: [Collections]
      parameters:
        - { $ref: '#/components/parameters/sitename' }
        - { $ref: '#/components/parameters/coll' }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: before, in: query, required: false, schema: { type: string }, description: cursor for the next page }
      responses:
        '200':
          description: Items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { type: object, properties: { id: { type: integer }, data: { description: the item as saved }, created_at: { type: string, format: date-time } } } }
                  next: { type: integer, nullable: true, description: cursor for the next page (pass as `before`) }
                  private: { type: boolean, description: present and true only for a private list }
        '403': { $ref: '#/components/responses/OriginMismatch' }
        '404':
          description: Site not found, or a private list and the caller is not its owner (code not_found)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: not found, code: not_found }
...
components:
  securitySchemes:
    apiKey: { type: apiKey, in: header, name: X-API-Key }
    connector:
      type: oauth2
      description: >-
        Per-person OAuth 2.1 (authorization code; PKCE S256 required for every self-registered
        client). An access token sent as `Authorization: Bearer` acts exactly as that person's
        X-API-Key on every route here. Discovery at /.well-known/oauth-authorization-server.
      flows:
        authorizationCode:
          authorizationUrl: https://simple-host.app/oauth/authorize
          tokenUrl: https://simple-host.app/oauth/token
          refreshUrl: https://simple-host.app/oauth/token
          scopes:
            sites: Publish and manage your sites and their data
  parameters:
...
    sitename: { name: sitename, in: path, required: true, schema: { type: string }, example: my-cool-site }
    coll: { name: coll, in: path, required: true, schema: { type: string }, example: signups }
    csrf: { name: X-SH-CSRF, in: header, required: false, schema: { type: string, enum: ['1'] }, description: Required on cookie-authenticated writes }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: What went wrong, in plain words. The wording may change; branch on `code`. }
        code:
          type: string
          description: >-
            A stable snake_case name for the refusal, present where one status
            has more than one meaning. site_exists (409: this account already
            has a site of that name), domain_taken (409: the address belongs to
            another site or person), invalid_name (400: not 1-63 lowercase
            letters, digits and hyphens), name_reserved (400: reserved by the
            platform), invalid_domain (400: not a domain that can be connected),
            site_quota_reached (403), append_only (409: public list items cannot
            be edited; the owner may still delete them), custom_domain_required (409: the site needs
            its own address first), not_an_object (409: the saved value is not a
            JSON object), private_visitor_only and private_needs_own_domain
            (403), use_custom_domain and visitor_auth_required (401),
            missing_api_key, wrong_auth_header (sent Authorization instead
            of X-API-Key) and invalid_api_key (unknown, revoked or signed-out
            key) (401), csrf_required (403), recently_deleted (409: the name
            belongs to a site in Recently deleted), site_not_deleted (409:
            restore of a live site), confirm_required (400),
            not_found (404, private collections), export_link_invalid (404:
            the download link expired or is not valid), site_suspended (403:
            the operator has taken the site down; changes and public reads of
            its data are refused until it is restored, `reason` says why),
            account_suspended (403: the operator has suspended this account;
            its keys, connected apps and sign-in are refused, `reason` says
            why). New codes may be added.
          example: domain_taken
  responses:
...
    TooLarge: { description: Payload too large, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    OriginMismatch: { description: Origin/Referer does not match the site, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...